Infinite row of vortices works!
In the v component equation I accidently used  sinh in the numerator
instead of sin. Also, by playing around with the spacing, a, I realized
having small 'a' would lead to only a few number of vortices being shown
in the graph with a wider radius. Once the spacing was increased the
graph looked similar to the finite row of vortices.

Having an even number of vortices would shift the 'o' off by 0.5.