Final Project literature and lesson 11 challenge task attempted
The stream lines and Cp was plotted where the format of lesson 10 was
followed. The difference was adding the vortex component.

Gamma is already calculated within the code so including gamma in the
class panel and calling it p.gamma was unnecessary. Also, the integral
function needs to be changed. The 0 and 1 or assigned "incorrectly" to
dxdz and dydz. However clarification is required because the way the
integral function is set up..the dxdz in the function placed where the
dydz is in the vortex integral.